add bold on navlink focus
