+++ src/APP/components/Header.jsx
@@ -1,24 +1,118 @@
+import { useState } from "react";
+import { Link, useLocation } from "react-router-dom";
+
+import menu from "../../assets/images/hamburger-menu.svg";
 import logo from "../../assets/images/sytLogo.png";
 
+const navLinks = [
+  {
+    id: 1,
+    link: "Home",
+    route: "/",
+  },
+  {
+    id: 2,
+    link: "About Us",
+    route: "/about-us",
+  },
+  {
+    id: 3,
+    link: "Community",
+    route: "/community",
+  },
+  {
+    id: 4,
+    link: "Products",
+    route: "/products",
+  },
+  {
+    id: 5,
+    link: "Blogs",
+    route: "/blogs",
+  },
+  {
+    id: 6,
+    link: "Resources",
+    route: "/resources",
+  },
+  // {
+  //   id: 7,
+  //   link: "Shop",
+  //   route: "/shop",
+  // },
+  // {
+  //   id: 8,
+  //   link: "Donate",
+  //   route: "/donate",
+  // },
+];
+
 function Header() {
+  const [showNavlinks, setShowNavlinks] = useState(false);
+
+  const { pathname } = useLocation();
+
   return (
-    <header className="hidden md:block text-gray-600 body-font">
-      <div className="container mx-auto flex flex-wrap p-5 flex-col md:flex-row items-center">
-        <a className="flex title-font font-medium items-center text-gray-900 mb-4 md:mb-0">
-          <img src={logo} className="h-24 mx-8" alt="Logo" />
-        </a>
-        <nav className="md:ml-auto flex flex-wrap items-center text-base justify-center">
-          <a className="mr-5 hover:text-gray-900">Home</a>
-          <a className="mr-5 hover:text-gray-900">About Us</a>
-          <a className="mr-5 hover:text-gray-900">Resources</a>
-          <a className="mr-5 hover:text-gray-900">Community</a>
-          <a className="mr-5 hover:text-gray-900">Products</a>
-          <a className="mr-5 hover:text-gray-900">Blog</a>
-          <a className="mr-5 hover:text-gray-900">Shop</a>
-          <a className="mr-5 hover:text-gray-900">Donate</a>
+    <div className="bg-[#252533] ">
+      <header className="py-2.5 md:px-10 px-5 flex items-center justify-between md:shadow-none shadow-md relative max-w-[1440px] md:mx-auto">
+        {/* logo */}
+        <Link to="/" className="md:w-[60px] w-12">
+          <img src={logo} alt="logo" className="w-full" />
+        </Link>
+
+        {/* mobile menu */}
+        <img
+          src={menu}
+          alt="logo"
+          className="md:hidden"
+          onClick={() => setShowNavlinks((prev) => !prev)}
+        />
+
+        {/* mobile navlinks */}
+        <nav
+          className="flex flex-col items-start justify-start gap-6 text-base absolute top-[70px] left-0 bg-gray-100 border-b w-full h-fit z-[1] p-5 pl-12"
+          style={
+            showNavlinks
+              ? { display: "flex", height: "calc(100vh - 65px)" }
+              : { display: "none" }
+          }
+        >
+          {navLinks.map(({ link, id, route }) => {
+            const isActive = pathname === route;
+            return (
+              <Link
+                key={id}
+                className={`${
+                  isActive ? "text-primary font-bold" : "text-gray-900"
+                } hover:text-primary transition-all duration-300 cursor-pointer`}
+                to={`${route}`}
+                onClick={() => setShowNavlinks(false)}
+              >
+                {link}
+              </Link>
+            );
+          })}
         </nav>
-      </div>
-    </header>
+
+        {/* navlinks */}
+        <nav className="md:flex hidden items-center gap-5 text-base">
+          {navLinks.map(({ id, link, route }) => {
+            const isActive = pathname === route;
+            return (
+              <Link
+                key={id}
+                className={`${
+                  isActive ? "text-primary font-bold" : "text-white"
+                } hover:text-primary transition-all duration-300 cursor-pointer`}
+                to={`${route}`}
+              >
+                {link}
+              </Link>
+            );
+          })}
+        </nav>
+      </header>
+    </div>
   );
 }
